extracting the examples in a file
