+++ app/package.json
@@ -29,7 +29,6 @@
     "react-dom": "^16.7.0-alpha.0",
     "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1",
-    "sanitize-html": "^1.19.1",
     "webpack-merge": "^4.1.4"
   },
   "devDependencies": {
